+++ servant-server/CHANGELOG.md
@@ -4,6 +4,9 @@
 * Add `err422` Unprocessable Entity
   ([#646](https://github.com/haskell-servant/servant/pull/646))
 
+* `Handler` is not abstract datatype. Migration hint: change `throwE` to `throwError`.
+  ([#641](https://github.com/haskell-servant/servant/issues/641))
+
 0.7.1
 ------